Factor out zulip into a reusable module.
